Added BOM without fasteners
I felt so bad after talking to everybody today and saying "I don't have  a BOM, but I'll have one soon." 
I don't want to wait to get you the most up to date info on the printer. So, here's everything but the screws and nuts. I'll add those eventually too, but for now this is the non-printable components that I can get to you the fastest from my hotel room. 

Thanks to everybody who showed up today at RMRFF!