Updated Toolhead support for endstop mod by rowokii
Updated the X_carriage_rear to accept the endstop mod from rowokii
This doesn't change anything for users who aren't using mechanical endstops, but makes it easier for people who do want to use mechanical switches to simply put them on.

Also did minor housekeeping for the assembly. It's still a mess, but it's a little bit better now.